Merged into ASP.NET Core Web API
"Dump20250927.sql" is current database file. Use it as the SQL document in MySQL. The "documents" table had to be renamed to "customer_records" due to ambiguous reference error with reserved keywords.

database.sql is there for reference purposes ONLY when creating the rest of the classes for the back end.